requirements.txt
python-telegram-bot==20.7
yt-dlp==2024.4.9
moviepy==1.0.3
pytz==2024.1
requests==2.31.0
